requirements.txt
fbapy==1.0.2
google-generativeai==0.3.2
python-dotenv==1.0.1
python-magic==0.4.27
